+++ tsup.config.ts
@@ -2,7 +2,8 @@ import { defineConfig } from "tsup"
 
 export default defineConfig({
   entry: {
-    index: "src/index.ts"
+    index: "src/index.ts",
+    constants: "src/constants.ts"
   },
   format: ["cjs", "esm"],
   dts: true,